Adjusted to classifier builders + static init
